PSEC-1483 WIP
