+++ inspectorInput/inspectorInput.go
@@ -1,14 +1,3 @@
 package inspectorInput
 
-type UserInput struct {
-	AwsAccount string
-	Usernname string
-	Region string
-	Profile string
-	Action string
-	FilterName string
-	FilterType string
-	ComparissionOperator string
-	MfaToken string
-}
